# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Interactive Fiction Platform API
  description: OpenAPI 3.0 Specification for Interactive Fiction Platform (DAG Novel System)
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
paths:
...
  /progress:
    post:
      summary: Record/Update reader current scene progress
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, novel_id, current_scene_id]
              properties:
                user_id:
                  type: integer
                  description: Required by the request validator; the authenticated JWT user ID is used when saving progress.
                novel_id:
                  type: integer
                current_scene_id:
                  type: integer
      responses:
        '201':
          description: Reading progress recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /history:
    get:
      summary: Get reader history list (Read-Only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Reader history list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [history]
                        properties:
                          history:
                            type: array
                            items:
                              $ref: '#/components/schemas/Novel'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Delete current user's reading history
      security:
        - BearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                novelIds:
                  type: array
                  items:
                    type: integer
      responses:
        '200':
          description: Reading history deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /user-endings:
    post:
      summary: Record unlocked ending for reader
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [novel_id, scene_id]
              properties:
                novel_id:
                  type: integer
                scene_id:
                  type: integer
      responses:
        '201':
          description: Ending unlocked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /bookshelves:
    get:
      summary: Get user bookshelf or check novel bookmark count
      parameters:
        - in: query
          name: user_id
          schema:
            type: integer
        - in: query
          name: novel_id
          schema:
            type: integer
      responses:
        '200':
          description: Bookshelf details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        oneOf:
                          - type: object
                            required: [bookshelf]
                            properties:
                              bookshelf:
                                type: array
                                items:
                                  $ref: '#/components/schemas/Novel'
                          - type: object
                            required: [novel_id, bookshelf_count]
                            properties:
                              novel_id:
                                type: integer
                              bookshelf_count:
                                type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Add novel to bookshelf (Bookmark)
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: novel_id
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              required: [novel_id]
              properties:
                novel_id:
                  type: integer
      responses:
        '201':
          description: Bookmarked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Remove novel from bookshelf
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: novel_id
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              required: [novel_id]
              properties:
                novel_id:
                  type: integer
      responses:
        '200':
          description: Bookmark removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /likes:
    post:
      summary: Like a novel
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [novel_id]
              properties:
                novel_id:
                  type: integer
      responses:
        '201':
          description: Novel liked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Unlike a novel
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: novel_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Novel unliked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /comments:
    post:
      summary: Add a comment to a novel
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [novel_id, content]
              properties:
                novel_id:
                  type: integer
                scene_id:
                  type: integer
                content:
                  type: string
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Remove a comment
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: comment_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Comment deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /choice-history:
    post:
      summary: Record reader choice history
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, choice_id]
              properties:
                user_id:
                  type: integer
                choice_id:
                  type: integer
      responses:
        '201':
          description: Choice history recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource Not Found Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    SuccessResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: integer
        message:
          type: string
        data:
          nullable: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: BAD_REQUEST
            message:
              type: string
              example: Invalid input payload
...
    Novel:
      type: object
      properties:
        novel_id:
          type: integer
        title:
          type: string
        captions:
          type: string
        introduction:
          type: string
        cover_image:
          type: string
        status:
          type: string
          enum: [draft, published, suspended]
        is_published:
          type: boolean
        is_completed:
          type: boolean
        is_banned:
          type: boolean
        ban_reason:
          type: string
        author_id:
          type: integer
        author_name:
          type: string
        pen_name:
          type: string
        views:
          type: integer
        chapter_count:
          type: integer
        scene_count:
          type: integer
        like_count:
          type: integer
        bookshelf_count:
          type: integer
...